renamed grabber to claw
